main.tf: remove interpolation
terraform v0.12.19 complains about the usage being deprecated,
and so should be modified. The complaint is:

Warning: Interpolation-only expressions are deprecated

  on .terraform/modules/ssh_config_update_host_entries/mcgrof-terraform-kdevops-add-host-ssh-config-ee46889/main.tf line 29, in resource "null_resource" "ssh_config_remove":
  29:     working_dir = "${path.module}"

Terraform 0.11 and earlier required all non-constant expressions to be
provided via interpolation syntax, but this pattern is now deprecated. To
silence this warning, remove the "${ sequence from the start and the }"
sequence from the end of this expression, leaving just the inner expression.

Template interpolation syntax is still used to construct strings from
expressions when the template includes multiple interpolation sequences or a
mixture of literal strings and interpolations. This deprecation applies only
to templates that consist entirely of a single interpolation sequence.

This should fix that.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>