main.tf: remove quoted reference on dependency
terraform v0.12.19 complains about a quoted reference now.

Warning: Quoted references are deprecated

  on .terraform/modules/ssh_config_update_host_entries/mcgrof-terraform-kdevops-add-host-ssh-config-ee46889/main.tf line 49, in resource "null_resource" "ssh_config_add":
  49:     "null_resource.ssh_config_remove",

In this context, references are expected literally rather than in quotes.
Terraform 0.11 and earlier required quotes, but quoted references are now
deprecated and will be removed in a future version of Terraform. Remove the
quotes surrounding this reference to silence this warning.

This should fix that.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>