variables.tf: remove quoted types
terraform v0.12.19 complains about a quoted type constraints.

Warning: Quoted type constraints are deprecated

  on .terraform/modules/ssh_config_update_host_entries/mcgrof-terraform-kdevops-add-host-ssh-config-ee46889/variables.tf line 2, in variable "ssh_config":
   2:   type = "string"

Terraform 0.11 and earlier required type constraints to be given in quotes,
but that form is now deprecated and will be removed in a future version of
Terraform. To silence this warning, remove the quotes around "string".

(and 10 more similar warnings elsewhere)

This should fix that.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>