main.tf: remove deprecated interpolation use
This will fix this interpolation deprecation complaint:

```
Warning: Interpolation-only expressions are deprecated

  on .terraform/modules/ssh_config_update_host_entries/terraform-kdevops-add-host-ssh-config-0.0.9/main.tf line 46, in resource "null_resource" "ssh_config_add":
  46:     working_dir = "${path.module}"

Terraform 0.11 and earlier required all non-constant expressions to be
provided via interpolation syntax, but this pattern is now deprecated. To
silence this warning, remove the "${ sequence from the start and the }"
sequence from the end of this expression, leaving just the inner expression.

Template interpolation syntax is still used to construct strings from
expressions when the template includes multiple interpolation sequences or a
mixture of literal strings and interpolations. This deprecation applies only
to templates that consist entirely of a single interpolation sequence.
```

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>