update_ssh_config.py: rely on python3
We're recommending the latest and greatest rolling distributions
for using kdevops. As such using python2 doesn't make sense.
But worst of all is just using /usr/bin/env python, given that

1) Some distributions are *not* recommending to use /usr/bin/env
2) Some distributions are not even providing a symlink for
/usr/bin/python

Since we *know* are recommending the latest and greatest for kdevops
base command and control center, just fixate on python3 for scripts
we run locally on the command and control center.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>