main.tf: simplify actions in one shot
The only reason that we were doing two separate calls before is that
we were being careful. But the python script now has a lot of tests,
and more can easily be added. We also currently already test the
ability to remove / add hosts in one shot, through the test:

  * test_0008_add_remove_hosts_one_shot_top()

Since we have built confidence on our testing, just do the operations
we need in one shot.

This also lets us explain the backup strategy for bring up method
consistantly.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>